Move components to profile folder
Update [username].tsx
