--- src/components/FreelancePetSitterProfile.tsx
@@ -1,12 +1,0 @@
-import Header from "./Header";
-
-const FreelancePetSitterProfile = (props: any) => {
-  return (
-    <div>
-      <Header></Header>
-      {props.username}
-    </div>
-  );
-};
-
-export default FreelancePetSitterProfile;
