+++ src/components/Profile/PetHotelProfile.tsx
@@ -1,4 +1,4 @@
-import Header from "./Header";
+import Header from "../Header";
 
 const PetHotelProfile = (props: any) => {
   return (
